Change at inventory statement controller

+++ iParty.Api/Views/Items/ItemView.cs
@@ -1,9 +1,6 @@
 ﻿using iParty.Api.Views.People;
 using iParty.Business.Models.Items;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace iParty.Api.Views.Items
 {
@@ -16,7 +13,7 @@ namespace iParty.Api.Views.Items
         public decimal Price { get; set; }
         public MeasurementUnit Unit { get; set; }
         public ProductOrService ProductOrService { get; set; }
-        public int AvailableQuantity { get; set; }
+        public decimal AvailableQuantity { get; set; }
         public RentOrSale ForRentOrSale { get; set; }
         public List<ScheduleView> Schedules { get; set; }
     }
